Updated files structure and class names and added dependent-gems workflow
