+++ .github/workflows/dependent-repos.json
@@ -1,0 +1,5 @@
+{
+  "repo": [
+    "unitsml/unitsml-ruby"
+  ]
+}
