+++ lib/unitsdb/dimensions.rb
@@ -1,13 +1,16 @@
 # frozen_string_literal: true
 
-require_relative "dimension"
+require_relative "dimensions/dimension"
 
 module Unitsdb
-  class Dimensions < Lutaml::Model::Serializable
-    attribute :dimension, Dimension, collection: true
+  class Dimensions
+    include Lutaml::Model::Serialize
+    model Config.model_for(:dimensions)
+
+    attribute :dimensions, Dimension, collection: true
 
     key_value do
-      map to: :dimension, root_mappings: {
+      map to: :dimensions, root_mappings: {
         id: :key
       }
     end
