--- lib/unitsdb/prefix_symbol.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-#   symbol:
-#     ascii: R
-#     html: R
-#     latex: R
-#     unicode: R
-
-module Unitsdb
-  class PrefixSymbol < Symbol
-  end
-end
