--- lib/unitsdb/root_unit.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-module Unitsdb
-  class RootUnit < Lutaml::Model::Serializable
-    attribute :unit, :string
-    attribute :power_denominator, :integer
-    attribute :power_numerator, :integer
-    attribute :prefix, :string
-
-    key_value do
-      map :unit, to: :unit
-      map :power_denominator, to: :power_denominator
-      map :power_numerator, to: :power_numerator
-      map :prefix, to: :prefix
-    end
-  end
-end
