--- lib/unitsdb/unit_symbol.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "symbol"
-
-module Unitsdb
-  class UnitSymbol < Symbol
-    attribute :id, :string
-    attribute :mathml, :string
-
-    key_value do
-      map :id, to: :id
-      map :mathml, to: :mathml
-    end
-  end
-end
