+++ spec/unitsdb/prefix_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Unitsdb::Prefix do
+RSpec.describe Unitsdb::Prefixes::Prefix do
   file_path = File.join(__dir__, "../fixtures/prefixes.yaml")
   prefixes_yaml = YAML.safe_load(IO.read(file_path))
